parsing, a few built in funtions + fork test
Fork has not been tested but i expect it too run

help opens a file a prints it to screen correctly (not tested on Linux server)
list directory and pause work too again not tested on Linux server

My parsing is working but input is broken down in a linear fashion.  the command a < b >> c would return
a will take b as input, b will appened to c. I need to figure out how to get a to append to c as that would be the program