now supports 1 way piping (update 14/03)
basicShell now supports piping 1 way as tested with the the 2 files uploaded.

initially, my parent was exec'ing the second program but this would cause my shell to quit as the shell (parent) was being overlaid.  Now I create 2 children if piping is required, they communicated as desired and then we return to the shell. 

next steps:
- fork of background exe
- implement batch made
- test,test and more tests