this is the most complete shell
it supports almost everything except batch mode and the more filter for help. possibility a few thinks i missed too 

I'm having problems with the batch mode. I can read the lines from file and assign them to a holding  buffer. but when I then try to move them to a array for evaluation, the contents of what i'm trying to copy seems to change.